+++ docs_src/sidebar/sidebar_data.py
@@ -42,6 +42,7 @@ sidebar_d = {
                 'vision.transform': '/vision.transform',
                 'vision.image': '/vision.image',
                 'vision.data': '/vision.data',
+                'vision.gan': '/vision.gan',
                 'vision.model overview': '/vision.models',
                 'vision.models.unet': '/vision.models.unet',
                 'vision.widgets': '/widgets'
@@ -81,7 +82,8 @@ sidebar_d = {
         'Custom ItemList': '/tutorial.itemlist',
     },
     'Doc authoring': {
-        'Overview': '/gen_doc',
+        'Instructions': '/gen_doc_main',
+        'gen_doc': '/gen_doc',
         'gen_doc.gen_notebooks': '/gen_doc.gen_notebooks',
         'gen_doc.nbdoc': '/gen_doc.nbdoc',
         'gen_doc.convert2html': '/gen_doc.convert2html',