+++ fastai/torch_core.py
@@ -56,6 +56,8 @@ fastai_types = {
     OptSplitFunc:'OptSplitFunc', PixelFunc:'PixelFunc', LightingFunc:'LightingFunc',
 }
 
+torch.set_num_threads(4) # OpenMP doesn't generally like too many threads
+
 bn_types = (nn.BatchNorm1d, nn.BatchNorm2d, nn.BatchNorm3d)
 defaults.device = torch.device('cuda') if torch.cuda.is_available() else torch.device('cpu')
 AdamW = partial(optim.Adam, betas=(0.9,0.99))
@@ -65,7 +67,11 @@ def tensor(x:Any, *rest)->Tensor:
     if len(rest): x = (x,)+rest
     # XXX: Pytorch bug in dataloader using num_workers>0; TODO: create repro and report
     if is_listy(x) and len(x)==0: return tensor(0)
-    return torch.tensor(x) if is_listy(x) else as_tensor(x)
+    res = torch.tensor(x) if is_listy(x) else as_tensor(x)
+    if res.dtype is torch.int32:
+        warn('Tensor is int32: upgrading to int64; for better performance use int64 input')
+        return res.long()
+    return res
 
 def np_address(x:np.ndarray)->int:
     "Address of `x` in memory."
@@ -74,7 +80,9 @@ def np_address(x:np.ndarray)->int:
 def to_detach(b:Tensors, cpu:bool=True):
     "Recursively detach lists of tensors in `b `; put them on the CPU if `cpu=True`."
     if is_listy(b): return [to_detach(o, cpu) for o in b]
-    return (b.detach().cpu() if cpu else b.detach()) if isinstance(b,Tensor) else b
+    if not isinstance(b,Tensor): return b
+    b = b.detach()
+    return b.cpu() if cpu else b
 
 def to_data(b:ItemsList):
     "Recursively map lists of items in `b ` to their wrapped data."